requirements.txt
easyocr==1.4.2
gpiozero==1.6.2
imutils==0.5.4
numpy==1.19.5
Pillow==9.1.0
pytesseract==0.3.9
torch==1.11.0
